+++ intellij-ext/modules/core/src/main/kotlin/com/metalbear/mirrord/MirrordLogger.kt
@@ -1,0 +1,7 @@
+package com.metalbear.mirrord
+
+import com.intellij.openapi.diagnostic.Logger
+
+object MirrordLogger {
+    public val logger = Logger.getInstance("mirrord")
+}