+++ mirrord/console/src/main.rs
@@ -93,7 +93,7 @@ async fn main() {
         .with(
             tracing_subscriber::fmt::layer()
                 .with_thread_ids(true)
-                .with_span_events(FmtSpan::ACTIVE)
+                .with_span_events(FmtSpan::NEW | FmtSpan::CLOSE)
                 .compact(),
         )
         .with(